# Change ddsketch mapping to improve performance.