--- a/pkg/logql/sketch/quantile.go
+++ b/pkg/logql/sketch/quantile.go
@@ -3,6 +3,7 @@ package sketch
 import (
 	"errors"
 	"fmt"
+	"sync"
 
 	"github.com/DataDog/sketches-go/ddsketch"
 	"github.com/DataDog/sketches-go/ddsketch/mapping"
@@ -18,6 +19,7 @@ type QuantileSketch interface {
 	Quantile(float64) (float64, error)
 	Merge(QuantileSketch) (QuantileSketch, error)
 	ToProto() *logproto.QuantileSketch
+	Release()
 }
 
 type QuantileSketchFactory func() QuantileSketch
@@ -42,9 +44,15 @@ type DDSketchQuantile struct {
 
 const relativeAccuracy = 0.01
 
+var ddsketchPool = sync.Pool{
+	New: func() any {
+		m, _ := mapping.NewCubicallyInterpolatedMapping(relativeAccuracy)
+		return ddsketch.NewDDSketchFromStoreProvider(m, store.DefaultProvider)
+	},
+}
+
 func NewDDSketch() *DDSketchQuantile {
-	m, _ := mapping.NewCubicallyInterpolatedMapping(relativeAccuracy)
-	s := ddsketch.NewDDSketchFromStoreProvider(m, store.DefaultProvider)
+	s := ddsketchPool.Get().(*ddsketch.DDSketch)
 	return &DDSketchQuantile{s}
 }
 
@@ -71,6 +79,11 @@ func (d *DDSketchQuantile) ToProto() *logproto.QuantileSketch {
 	return &logproto.QuantileSketch{
 		Sketch: sketch,
 	}
+}
+
+func (d *DDSketchQuantile) Release() {
+	d.DDSketch.Clear()
+	ddsketchPool.Put(d.DDSketch)
 }
 
 func DDSketchQuantileFromProto(buf []byte) (*DDSketchQuantile, error) {
@@ -131,6 +144,8 @@ func (d *TDigestQuantile) ToProto() *logproto.QuantileSketch {
 		},
 	}
 }
+
+func (d *TDigestQuantile) Release() {}
 
 func TDigestQuantileFromProto(proto *logproto.TDigest) *TDigestQuantile {
 	q := &TDigestQuantile{tdigest.NewWithCompression(proto.Compression)}