@cstyan 2024-01-02
What are the three different benchmark names? Cubic and pool are the sketch mapping change and then the sketch mapping change + pooling, but what is the sort?