@jeschkies 2024-01-03
> What are the three different benchmark names?

- logarithmic.log: LogarithmicMapping
- cubic.log: CubicallyInterpolatedMapping
- sort.log: CubicallyInterpolatedMapping + https://github.com/DataDog/sketches-go/pull/76         - pool.log: CubicallyInterpolatedMapping + https://github.com/DataDog/sketches-go/pull/76 + DDSketch pooling